requirement.txt
absl-py==1.1.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses==0.8
decorator==4.4.2
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py @ file:///D:/bld/h5py_1617647698074/work
idna==3.3
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
joblib==1.1.1
kiwisolver @ file:///D:/bld/kiwisolver_1610099971815/work
Markdown==3.3.7
matplotlib==3.1.2
networkx==2.5.1
numpy==1.17.0
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.1.2.30
pandas==1.1.5
Pillow==8.4.0
prefetch-generator==1.0.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline @ file:///D:/bld/pyreadline_1632466759315/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.6
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.2.1
seaborn==0.11.2
Shapely==1.8.5.post1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado @ file:///D:/bld/tornado_1610094881553/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1617680194207/work
typing_extensions==4.1.1
urllib3==1.26.10
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
